pyproject.toml
[tool.poetry]
name = "drf-project"
version = "0.1.0"
description = ""
authors = ["Nikolay Averin <rizzzon116@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
djangorestframework = "^3.15.1"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
pillow = "^10.3.0"
django-phonenumber-field = "^7.3.0"
black = "^24.4.2"
ipython = "^8.25.0"
isort = "^5.13.2"
django-filter = "^24.2"
djangorestframework-simplejwt = "^5.3.1"
coverage = "^7.5.4"
drf-yasg = "^1.21.7"
stripe = "^10.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
